refactor: consolidate cross-cutting types and rename useTimer to named export
Move `Palette`, `ToastVariant`, and `IDialogAction` to `src/internal/types.ts`
alongside the existing `IBaseProps` / `IFieldProps` home. Components and signals
now depend on a shared internal types module rather than each other — `@signals/*`
no longer imports enums from `@components/*`, and `@internal/DialogBase` no longer
owns a type consumed from `@signals/dialogs`.

Public enums (`Palette`, `ToastVariant`) are re-exported from `@components/Theme`
and `@components/Toast` so `@auldrant/ui/components` consumers see no change.

Rename `useTimer` from default to named export in `src/hooks.ts` for consistency
with the file's other exports (`usePage`, `page`, `ITimerControls`). Consumers
update: `import useTimer from '@auldrant/ui/hooks'` → `import { useTimer } ...`.

Also fix `src/components/index.ts` to use the `@styles/...` alias instead of
`../styles/...` (per project import rules).
