+++ src/components/Toast.tsx
@@ -1,19 +1,14 @@
 import Icon, { IconName } from '@components/Icon';
 import Tooltip from '@components/Tooltip';
-import useTimer from '@hooks';
+import { useTimer } from '@hooks';
+import { ToastVariant } from '@internal/types';
 import type { IBaseProps } from '@internal/types';
 import { useSignal } from '@preact/signals';
 import styles from '@styles/Toast.module.css';
 import { cx } from '@utils';
 import type { FunctionComponent } from 'preact';
 
-/** Severity variants for {@link Toast}. Controls color and leading icon. */
-export enum ToastVariant {
-	info = 'info',
-	success = 'success',
-	warning = 'warning',
-	error = 'error',
-}
+export { ToastVariant } from '@internal/types';
 
 /** Props for {@link Toast}. */
 interface IToastProps extends IBaseProps {
